Enhance auth and product services with updated validations, soft deletion, and robust entity behaviors

+++ authService/app/src/main/java/org/example/repos/RefreshTokenRepo.java
@@ -8,7 +8,7 @@ import org.springframework.stereotype.Repository;
 import java.util.Optional;
 
 @Repository
-public interface RefreshTokenRepo extends JpaRepository<RefreshToken, Integer> {
+public interface RefreshTokenRepo extends JpaRepository<RefreshToken, Long> {
 
     Optional<RefreshToken> findByToken(String token);
